+++ examples/generative/stormcast/train.py
@@ -24,7 +24,6 @@ import wandb
 import glob
 from omegaconf import DictConfig, OmegaConf
 from physicsnemo.distributed import DistributedManager
-from physicsnemo.launch.logging import PythonLogger, RankZeroLoggingWrapper
 
 from utils.trainer import training_loop
 
@@ -36,8 +35,6 @@ def main(cfg: DictConfig) -> None:
     # Initialize
     DistributedManager.initialize()
     dist = DistributedManager()
-    logger = PythonLogger("main")
-    logger0 = RankZeroLoggingWrapper(logger, dist)  # Log only from rank 0
 
     # Random seed.
     if cfg.training.seed < 0:
@@ -45,31 +42,25 @@ def main(cfg: DictConfig) -> None:
         torch.distributed.broadcast(seed, src=0)
         cfg.training.seed = int(seed)
 
-    # Resume from specified checkpoint, if provided
-    if cfg.training.resume_checkpoint is not None:
-        resume = cfg.training.resume_checkpoint
-        if not os.path.isfile(resume) or not resume.endswith(".pt"):
+    # Start from specified checkpoint, if provided
+    if cfg.training.initial_weights is not None:
+        weights_path = cfg.training.initial_weights
+        if not os.path.isfile(weights_path) or not (
+            weights_path.endswith(".mdlus") or weights_path.endswith(".pt")
+        ):
             raise ValueError(
-                "training.resume_checkpoint must point to a physicsnemo .pt checkpoint from a previous training run"
+                "training.initial_weights must point to a physicsnemo .mdlus or .pt checkpoint from a previous training run"
             )
 
-    # If run directory already exists, then resume training from last checkpoint
+    # If checkpoint directory already exists, then resume training from last checkpoint
     wandb_resume = False
-    if os.path.exists(cfg.training.rundir):
-        training_states = sorted(
-            glob.glob(os.path.join(cfg.training.rundir, "checkpoints/checkpoint*.pt"))
-        )
-        if training_states:
-            logger0.info(
-                "Resuming training from previous run_dir: " + cfg.training.rundir
-            )
-            last_training_state = training_states[-1]
-            cfg.training.resume_checkpoint = last_training_state
-            logger0.info(
-                "Resuming training from previous checkpoint file: "
-                + last_training_state
-            )
-            wandb_resume = True
+    os.makedirs(cfg.training.rundir, exist_ok=True)
+    net_name = "regression" if cfg.training.loss == "regression" else "diffusion"
+    training_states = glob.glob(
+        os.path.join(cfg.training.rundir, f"checkpoints_{net_name}/checkpoint*.pt")
+    )
+    if training_states:
+        wandb_resume = True
 
     # Setup wandb, if enabled
     if dist.rank == 0 and cfg.training.log_to_wandb:
@@ -81,7 +72,7 @@ def main(cfg: DictConfig) -> None:
             project=project,
             entity=entity,
             resume=wandb_resume,
-            mode="online",
+            mode=cfg.training.wandb_mode,
         )
 
     # Train.