+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simplified CorrDiff config files, updated default values
 - Refactored CorrDiff losses and samplers to use the patching API
 - Support for non-square images and patches in patch-based diffusion
+- Support for training custom StormCast models and various other improvements for StormCast
 
 ### Deprecated